backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
scapy
google-generativeai
reportlab
python-dateutil
python-jose
passlib[bcrypt]
bcrypt==4.0.1
requests
bson
pydantic
serial